+++ packages/editor/js/extensions/libs/layout/styles.js
@@ -139,6 +139,11 @@ export const LayoutStyles = ({
 				),
 			});
 
+			const alignProp: string =
+				_attributes?.blockeraFlexLayout?.direction === 'column'
+					? 'justify-content'
+					: 'align-items';
+
 			styleGroup.push({
 				selector: pickedSelector,
 				declarations: computedCssDeclarations(
@@ -147,7 +152,7 @@ export const LayoutStyles = ({
 							{
 								...staticDefinitionParams,
 								properties: {
-									'align-items':
+									[alignProp]:
 										_attributes.blockeraFlexLayout
 											.alignItems,
 								},
@@ -169,6 +174,11 @@ export const LayoutStyles = ({
 				),
 			});
 
+			const justifyProp: string =
+				_attributes?.blockeraFlexLayout?.direction === 'column'
+					? 'align-items'
+					: 'justify-content';
+
 			styleGroup.push({
 				selector: pickedSelector,
 				declarations: computedCssDeclarations(
@@ -177,7 +187,7 @@ export const LayoutStyles = ({
 							{
 								...staticDefinitionParams,
 								properties: {
-									'justify-content':
+									[justifyProp]:
 										_attributes.blockeraFlexLayout
 											.justifyContent,
 								},